+++ packages/webpack/src/plugins/togglePointInjection/setupSchemeModules/index.js
@@ -1,5 +1,5 @@
 import { PLUGIN_NAME, POINT_CUTS, JOIN_POINTS, SCHEME } from "../constants.js";
-import generateJoinPoint from "./generateJoinPoint.js";
+import generateJoinPoint from "./generateJoinPoint/index.js";
 import generatePointCut from "./generatePointCut.js";
 
 const setupSchemeModules = ({
@@ -11,13 +11,13 @@ const setupSchemeModules = ({
   NormalModule.getCompilationHooks(compilation)
     .readResource.for(SCHEME)
     .tap(PLUGIN_NAME, ({ resourcePath }) => {
-      const [, type, path] = resourcePath.split(/:(.*?):(.*)/, 3);
+      const [, type, joinPointPath] = resourcePath.split(/:(.*?):(.*)/, 3);
       switch (type) {
         case POINT_CUTS: {
-          return generatePointCut({ pointCuts, path });
+          return generatePointCut({ pointCuts, joinPointPath });
         }
         case JOIN_POINTS: {
-          return generateJoinPoint({ joinPointFiles, path });
+          return generateJoinPoint({ joinPointFiles, joinPointPath });
         }
       }
     });