+++ packages/webpack/src/toggleHandlers/pathSegmentToggleHandler.test.js
@@ -5,9 +5,7 @@ const togglePoint = jest.fn(() => toggleOutcome);
 const joinPoint = Symbol("mock-join-point");
 
 describe("pathSegmentToggleHandler", () => {
-  let result, variantsMap;
-  const variants = (key) => variantsMap[key];
-  variants.keys = () => Object.keys(variantsMap);
+  let result;
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -17,15 +15,25 @@ describe("pathSegmentToggleHandler", () => {
     const keyArray = [...Array(segmentCount).keys()];
 
     describe(`given a list of variant paths with ${segmentCount} path segments (after the variants path)`, () => {
+      let variantPathMap;
+
       beforeEach(() => {
         const segments = keyArray.map((key) => `test-segment-${key}/`);
-        variantsMap = {
-          [`./__variants__/${segments.join("")}test-variant.js`]:
-            Symbol("test-variant"),
-          [`./__variants__/${segments.reverse().join("")}test-variant.js`]:
+        variantPathMap = new Map([
+          [
+            `./__variants__/${segments.join("")}test-variant.js`,
             Symbol("test-variant")
-        };
-        result = pathSegmentToggleHandler({ togglePoint, joinPoint, variants });
+          ],
+          [
+            `./__variants__/${segments.reverse().join("")}test-variant.js`,
+            Symbol("test-variant")
+          ]
+        ]);
+        result = pathSegmentToggleHandler({
+          togglePoint,
+          joinPoint,
+          variantPathMap
+        });
       });
 
       it("should call the toggle point with the join point module and a map", () => {
@@ -44,14 +52,14 @@ describe("pathSegmentToggleHandler", () => {
         });
 
         it("should return a map containing maps for each segment, concluding with the variant at the leaf node", () => {
-          for (const key of Object.keys(variantsMap)) {
+          for (const key of Object.keys(variantPathMap)) {
             const segments = key.split("/").slice(0, -1);
             let node = map;
             for (const segment of segments.slice(2)) {
               expect(node.has(segment)).toBe(true);
               node = node.get(segment);
             }
-            expect(node).toBe(variantsMap[key]);
+            expect(node).toBe(variantPathMap[key]);
           }
         });
       });