+++ examples/express/docs/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.6] - 2025-07-29
+
+### Fixed
+
+- removed "Vary" header from "animals" example, the page is meant to be un-cacheable, and the value was wrong in any case
+
 ## [0.2.5] - 2025-05-27
 
 ### Changed